+++ src/App.js
@@ -1,6 +1,7 @@
 import React from 'react';
 import { Route } from 'react-router-dom';
 import { Home, Auth } from 'pages';
+import Problem from 'pages/Problem';
 
 // 라우트 정의 컴포넌트
 
@@ -9,6 +10,7 @@ function App() {
     <div>
       <Route exact path="/" component={Home}/>
       <Route path="/auth" component={Auth}/>
+      <Route path="/list" component={Problem}/>
     </div>
   );
 }
