+++ src/containers/problem/back/BackList.jsx
@@ -1,0 +1,14 @@
+import React, { Component } from 'react';
+
+
+class BackList extends Component {
+    render () {
+        return (
+            <div>
+                <p>백리스트</p>
+            </div>
+        );
+    }
+}
+
+export default BackList;
