+++ src/containers/problem/front/FrontList.jsx
@@ -1,0 +1,14 @@
+import React, { Component } from 'react';
+
+
+class FrontList extends Component {
+    render () {
+        return (
+            <div>
+                <p>프론트리스트</p>
+            </div>
+        );
+    }
+}
+
+export default FrontList;
